Remove dir type related method from relative path
